Refactor server
